# Conflicts:
#	ARIS.xcodeproj/project.pbxproj
#	Info.plist
#	src/tab_controllers/map/MapViewController.m

+++ src/data_objects/Trigger.h
@@ -29,8 +29,8 @@
   BOOL hidden;
   BOOL trigger_on_enter;
   NSString *qr_code;
-  int seconds;
-  int time_left; //client only!
+  long seconds;
+  long time_left; //client only!
 }
 
 @property (nonatomic, assign) long trigger_id;
@@ -49,8 +49,8 @@
 @property (nonatomic, assign) BOOL hidden;
 @property (nonatomic, assign) BOOL trigger_on_enter;
 @property (nonatomic, copy)   NSString *qr_code;
-@property (nonatomic, assign) int seconds;
-@property (nonatomic, assign) int time_left;
+@property (nonatomic, assign) long seconds;
+@property (nonatomic, assign) long time_left;
 
 - (id) initWithDictionary:(NSDictionary *)dict;
 - (NSString *) serialize;
